// src/app/models/ods.interface.ts
export interface Ods{

    id : number;
    fin_pobreza : boolean;
    hambre_cero: boolean;
    salud_bienestar: boolean;
    educacion_calidad: boolean;
    igualdad_genero : boolean;
    agua_limp_sanea : boolean;
    enr_ase_nocon: boolean;
    trab_dec_creeco: boolean;
    ind_ino_inf: boolean;
    reduc_desig: boolean;
    ciu_com_sos: boolean;
    prod_con_res: boolean;
    acc_por_cli: boolean;
    vida_sub: boolean;
    vida_eco_terr: boolean;
    paz_jus_instsol: boolean;
    ali_lograr_obj: boolean;
    ecoId:number;
    ecoCircularId:number;
    fecha_registro: Date;
    fecha_actualizacion: Date;
}